Praktikum 5

+++ app/Http/Controllers/ProductController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Product;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate; 
 
 class ProductController extends Controller
 {
@@ -44,6 +45,8 @@ class ProductController extends Controller
     {
         $product = Product::findOrFail($id);
 
+        Gate::authorize('update', $product);
+
         $validated = $request->validate([
             'name' => 'sometimes|string|max:255',
             'qty' => 'sometimes|integer',
@@ -58,6 +61,8 @@ class ProductController extends Controller
 
     public function edit(Product $product)
     {
+        Gate::authorize('update', $product);
+
         $users = User::orderBy('name')->get();
         return view('product.edit', compact('product', 'users'));
     }
@@ -65,6 +70,9 @@ class ProductController extends Controller
     public function delete($id)
     {
         $product = Product::findOrFail($id);
+
+        Gate::authorize('delete', $product);
+
         $product->delete();
 
         return redirect()->route('product.index')->with('success', 'Product berhasil dihapus');
